Added model comments, create new models, fix factories path

+++ database/migrations/2020_05_15_131656_description_profile.php
@@ -16,7 +16,7 @@ class DescriptionProfile extends Migration
         Schema::create('description_profile', function (Blueprint $table) {
             $table->Increments('description_profile_id');
             $table->integer('id_user')->comment('user id, from the users table ');
-            $table->integer('lvl')->comment('User level')->default(1);
+            $table->smallInteger('lvl')->comment('User level')->default(1);
             $table->float('exp')->comment('Experience');
             $table->string('real_name', 100)->nullable();
             $table->date('date_born')->nullable();
